Refactor product service: update imports, remove unused stack, and add mock products
