+++ lib/product-service/handlers/mock-products.ts
